+++ webpack/prod.js
@@ -4,22 +4,12 @@
 'use strict';
 
 var baseConfig = require('./base'),
-    _ = require('lodash'),
     ExtractTextPlugin = require('extract-text-webpack-plugin'),
-    HtmlWebpackPlugin = require('html-webpack-plugin'),
     webpack = require('webpack');
 
-module.exports = _.merge({}, baseConfig, {
-    plugins: [
-        new ExtractTextPlugin('[hash].css'),
-        new HtmlWebpackPlugin({
-            template: 'html?interpolate!./src/index.ejs',
-        }),
-        new webpack.optimize.UglifyJsPlugin({
-            minimize: true,
-        }),
-    ],
-});
+baseConfig.plugins.push(new webpack.optimize.UglifyJsPlugin({
+    minimize: true,
+}));
 
 // Apply Extractor to every css loader
 baseConfig.module.loaders = baseConfig.module.loaders.map((loader) => {